Moved the implementation classes to an `internal` package so these implementations are not leaked by OSGI.
