--- components/ledger/ledger-utxo-token-cache/src/main/kotlin/net/corda/ledger/utxo/token/cache/entities/TokenPoolCacheImpl.kt
@@ -1,10 +1,0 @@
-package net.corda.ledger.utxo.token.cache.entities
-
-class TokenPoolCacheImpl : TokenPoolCache {
-
-    private val cache = mutableMapOf<TokenPoolKey, TokenCache>()
-
-    override fun get(poolKey: TokenPoolKey): TokenCache {
-        return cache.getOrPut(poolKey) { TokenCacheImpl() }
-    }
-}
