+++ components/ledger/ledger-utxo-token-cache/src/main/kotlin/net/corda/ledger/utxo/token/cache/factories/TokenCacheComponentFactory.kt
@@ -6,7 +6,7 @@ import net.corda.flow.external.events.responses.factory.ExternalEventResponseFac
 import net.corda.ledger.utxo.token.cache.converters.EntityConverterImpl
 import net.corda.ledger.utxo.token.cache.converters.EventConverterImpl
 import net.corda.ledger.utxo.token.cache.entities.TokenEvent
-import net.corda.ledger.utxo.token.cache.entities.TokenPoolCacheImpl
+import net.corda.ledger.utxo.token.cache.entities.internal.TokenPoolCacheImpl
 import net.corda.ledger.utxo.token.cache.handlers.TokenBalanceQueryEventHandler
 import net.corda.ledger.utxo.token.cache.handlers.TokenClaimQueryEventHandler
 import net.corda.ledger.utxo.token.cache.handlers.TokenClaimReleaseEventHandler
